requirements.txt
crewai
crewai-tools
requests
python-dotenv
langchain-google-genai
